fastapi/requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
bcrypt==4.0.1
black==26.3.1
cryptography==46.0.5
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.1
httptools==0.7.1
locust==2.43.3
passlib==1.7.4
pipdeptree==2.31.0
pre_commit==4.5.1
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pymongo==4.16.0
pyotp==2.9.0
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-jose==3.5.0
python-multipart==0.0.22
ruff==0.15.6
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0


# pipdeptree --warn silence --depth 0
# alembic revision --autogenerate -m ""
# alembic upgrade head